+++ apollo-ios-pagination/Sources/ApolloPagination/GraphQLQueryPager.swift
@@ -1,28 +1,31 @@
 import Apollo
 import ApolloAPI
 import Combine
+import Dispatch
 
 /// Type-erases a query pager, transforming data from a generic type to a specific type, often a view model or array of view models.
-public class AnyAsyncGraphQLQueryPager<Model> {
+public class GraphQLQueryPager<Model> {
   public typealias Output = Result<(Model, UpdateSource), Error>
   private let _subject: CurrentValueSubject<Output?, Never> = .init(nil)
-  public var publisher: AnyPublisher<Output, Never> { _subject.compactMap({ $0 }).eraseToAnyPublisher() }
-  public var cancellables = [AnyCancellable]()
-  public let pager: any AsyncPagerType
 
-  public var canLoadNext: Bool { get async { await pager.canLoadNext } }
-  public var canLoadPrevious: Bool { get async { await pager.canLoadPrevious } }
+  /// The `publisher` is the intended access point for using the pager as a `Combine` stream.
+  public var publisher: AnyPublisher<Output, Never> { _subject.compactMap { $0 }.eraseToAnyPublisher() }
+  public var cancellables = [AnyCancellable]()
+  public let pager: any PagerType
+
+  public var canLoadNext: Bool { pager.canLoadNext }
+  public var canLoadPrevious: Bool { pager.canLoadPrevious }
 
   /// Type-erases a given pager, transforming data to a model as pagination receives new results.
   /// - Parameters:
   ///   - pager: Pager to type-erase.
   ///   - transform: Transformation from an initial page and array of paginated pages to a given view model.
-  public init<Pager: AsyncGraphQLQueryPager<InitialQuery, NextQuery>, InitialQuery, NextQuery>(
+  public init<Pager: GraphQLQueryPagerCoordinator<InitialQuery, NextQuery>, InitialQuery, NextQuery>(
     pager: Pager,
     transform: @escaping ([NextQuery.Data], InitialQuery.Data, [NextQuery.Data]) throws -> Model
-  ) async {
+  ) {
     self.pager = pager
-    await pager.subscribe { [weak self] result in
+    pager.subscribe { [weak self] result in
       guard let self else { return }
       let returnValue: Output
 
@@ -39,7 +42,7 @@ public class AnyAsyncGraphQLQueryPager<Model> {
       }
 
       _subject.send(returnValue)
-    }.store(in: &cancellables)
+    }
   }
 
   /// Type-erases a given pager, transforming the initial page to an array of models, and the
@@ -49,7 +52,7 @@ public class AnyAsyncGraphQLQueryPager<Model> {
   ///   - initialTransform: Initial transformation from the initial page to an array of models.
   ///   - nextPageTransform: Transformation to execute on each subseqent page to an array of models.
   public convenience init<
-    Pager: AsyncGraphQLQueryPager<InitialQuery, NextQuery>,
+    Pager: GraphQLQueryPagerCoordinator<InitialQuery, NextQuery>,
     InitialQuery,
     NextQuery,
     Element
@@ -57,8 +60,8 @@ public class AnyAsyncGraphQLQueryPager<Model> {
     pager: Pager,
     initialTransform: @escaping (InitialQuery.Data) throws -> Model,
     pageTransform: @escaping (NextQuery.Data) throws -> Model
-  ) async where Model: RangeReplaceableCollection, Model.Element == Element {
-    await self.init(
+  ) where Model: RangeReplaceableCollection, Model.Element == Element {
+    self.init(
       pager: pager,
       transform: { previousData, initialData, nextData in
         let previous = try previousData.flatMap { try pageTransform($0) }
@@ -69,9 +72,13 @@ public class AnyAsyncGraphQLQueryPager<Model> {
     )
   }
 
+  deinit {
+    pager.cancel()
+  }
+
   /// Subscribe to the results of the pager, with the management of the subscriber being stored internally to the `AnyGraphQLQueryPager`.
-  /// - Parameter completion: The closure to trigger when new values come in.
-  public func subscribe(completion: @MainActor @escaping (Output) -> Void) {
+  /// - Parameter completion: The closure to trigger when new values come in. Guaranteed to run on the main thread.
+  public func subscribe(completion: @escaping @MainActor (Output) -> Void) {
     publisher.sink { result in
       Task { await completion(result) }
     }.store(in: &cancellables)
@@ -80,72 +87,81 @@ public class AnyAsyncGraphQLQueryPager<Model> {
   /// Load the next page, if available.
   /// - Parameters:
   ///   - cachePolicy: The Apollo `CachePolicy` to use. Defaults to `returnCacheDataAndFetch`.
+  ///   - callbackQueue: The `DispatchQueue` that the `completion` fires on. Defaults to `main`.
+  ///   - completion: A completion block that will always trigger after the execution of this operation. Passes an optional error, of type `PaginationError`, if there was an internal error related to pagination. Does not surface network errors. Defaults to `nil`.
   public func loadNext(
-    cachePolicy: CachePolicy = .returnCacheDataAndFetch
-  ) async throws {
-    try await pager.loadNext(cachePolicy: cachePolicy)
+    cachePolicy: CachePolicy = .returnCacheDataAndFetch,
+    callbackQueue: DispatchQueue = .main,
+    completion: ((PaginationError?) -> Void)? = nil
+  ) {
+    pager.loadNext(cachePolicy: cachePolicy, callbackQueue: callbackQueue, completion: completion)
   }
 
   /// Load the previous page, if available.
   /// - Parameters:
   ///   - cachePolicy: The Apollo `CachePolicy` to use. Defaults to `returnCacheDataAndFetch`.
+  ///   - callbackQueue: The `DispatchQueue` that the `completion` fires on. Defaults to `main`.
+  ///   - completion: A completion block that will always trigger after the execution of this operation. Passes an optional error, of type `PaginationError`, if there was an internal error related to pagination. Does not surface network errors. Defaults to `nil`.
   public func loadPrevious(
-    cachePolicy: CachePolicy = .returnCacheDataAndFetch
-  ) async throws {
-    try await pager.loadPrevious(cachePolicy: cachePolicy)
+    cachePolicy: CachePolicy = .returnCacheDataAndFetch,
+    callbackQueue: DispatchQueue = .main,
+    completion: ((PaginationError?) -> Void)? = nil
+  ) {
+    pager.loadPrevious(cachePolicy: cachePolicy, callbackQueue: callbackQueue, completion: completion)
   }
 
   /// Loads all pages.
   /// - Parameters:
   ///   - fetchFromInitialPage: Pass true to begin loading from the initial page; otherwise pass false.  Defaults to `true`.  **NOTE**: Loading all pages with this value set to `false` requires that the initial page has already been loaded previously.
+  ///   - callbackQueue: The `DispatchQueue` that the `completion` fires on. Defaults to `main`.
+  ///   - completion: A completion block that will always trigger after the execution of this operation. Passes an optional error, of type `PaginationError`, if there was an internal error related to pagination. Does not surface network errors. Defaults to `nil`.
   public func loadAll(
-    fetchFromInitialPage: Bool = true
-  ) async throws {
-    try await pager.loadAll(fetchFromInitialPage: fetchFromInitialPage)
+    fetchFromInitialPage: Bool = true,
+    callbackQueue: DispatchQueue = .main,
+    completion: ((PaginationError?) -> Void)? = nil
+  ) {
+    pager.loadAll(fetchFromInitialPage: fetchFromInitialPage, callbackQueue: callbackQueue, completion: completion)
   }
 
   /// Discards pagination state and fetches the first page from scratch.
   /// - Parameter cachePolicy: The apollo cache policy to trigger the first fetch with. Defaults to `fetchIgnoringCacheData`.
-  public func refetch(cachePolicy: CachePolicy = .fetchIgnoringCacheData) async {
-    await pager.refetch(cachePolicy: cachePolicy)
+  public func refetch(cachePolicy: CachePolicy = .fetchIgnoringCacheData) {
+    pager.refetch(cachePolicy: cachePolicy)
   }
 
   /// Fetches the first page.
-  public func fetch() async {
-    await pager.fetch()
+  public func fetch() {
+    pager.fetch()
   }
 
   /// Resets pagination state and cancels further updates from the pager.
-  public func cancel() async {
-    await pager.cancel()
+  public func cancel() {
+    pager.cancel()
   }
 }
 
-extension AsyncGraphQLQueryPager {
-  nonisolated func eraseToAnyPager<T>(
+public extension GraphQLQueryPagerCoordinator {
+  func eraseToAnyPager<T>(
     transform: @escaping ([PaginatedQuery.Data], InitialQuery.Data, [PaginatedQuery.Data]) throws -> T
-  ) async -> AnyAsyncGraphQLQueryPager<T> {
-    await AnyAsyncGraphQLQueryPager(
-      pager: self,
-      transform: transform
-    )
+  ) -> GraphQLQueryPager<T> {
+    GraphQLQueryPager(pager: self, transform: transform)
   }
 
-  nonisolated func eraseToAnyPager<T, S: RangeReplaceableCollection>(
+  func eraseToAnyPager<T, S: RangeReplaceableCollection>(
     initialTransform: @escaping (InitialQuery.Data) throws -> S,
-    pageTransform: @escaping (PaginatedQuery.Data) throws -> S
-  ) async -> AnyAsyncGraphQLQueryPager<S> where T == S.Element {
-    await AnyAsyncGraphQLQueryPager(
+    nextPageTransform: @escaping (PaginatedQuery.Data) throws -> S
+  ) -> GraphQLQueryPager<S> where T == S.Element {
+    GraphQLQueryPager(
       pager: self,
       initialTransform: initialTransform,
-      pageTransform: pageTransform
+      pageTransform: nextPageTransform
     )
   }
 
-  nonisolated func eraseToAnyPager<T, S: RangeReplaceableCollection>(
+  func eraseToAnyPager<T, S: RangeReplaceableCollection>(
     transform: @escaping (InitialQuery.Data) throws -> S
-  ) async -> AnyAsyncGraphQLQueryPager<S> where InitialQuery == PaginatedQuery, T == S.Element {
-    await AnyAsyncGraphQLQueryPager(
+  ) -> GraphQLQueryPager<S> where InitialQuery == PaginatedQuery, T == S.Element {
+    GraphQLQueryPager(
       pager: self,
       initialTransform: transform,
       pageTransform: transform
